Removed StackUtils.cs and implemented them in seperate classes. Added BalancedParenthesesChecker.cs, NextGreaterElement.cs and ReverseStack.cs
