+++ README.md
@@ -242,7 +242,9 @@ find more than one implementation for the same objective but using different alg
     * [Array-based Stack](./DataStructures/Stack/ArrayBasedStack.cs)
     * [List-based Stack](./DataStructures/Stack/ListBasedStack.cs)
     * [Queue-based Stack](./DataStructures/Stack/QueueBasedStack.cs)
-    * [Stack Utils](./DataStructures/Stack/StackUtils.cs)
+    * [Next Greater Element](./DataStructures/Stack/NextGreaterElement.cs)
+    * [BalancedParenthesesChecker](./DataStructures/Stack/BalancedParenthesesChecker.cs)
+    * [Reverse Stack](./DataStructures/Stack/ReverseStack.cs)
   * [Heap](./DataStructures/Heap)
     * [Min-Max Heap](./DataStructures/Heap/MinMaxHeap.cs)
     * [Binary Heap](./DataStructures/Heap/BinaryHeap.cs)
